Added Clumped + major refactor
Moved uniform module to chunked, since uniformly chunked decorators are
indeed a special instance of chunked.

Added a Clumped decorator, which is a special version of Chunked that
compactly represents consecutive offsets with equal strides.
To enable this, ClumpedOffsets were added, thus reusing much of what was
already implemented in Chunked.

Added a GetOffset trait to share an interface between Offsets and
ClumpedOffsets. Other types of offsets may be needed later.