Gradually migrating to bytemuck to improve soundness
bytemuck provides the very strict Pod trait which can be employed in
UniChunked types, since those are typically used for Pod types anyways.
Although Pod itself is strict, if a type T is not Pod, a collection of T
can still be uniformly chunked using ChunkedN, which avoids transmutation.
So even though this change restricts the usage of Chunked# types, it
doesn't take away from the capabilities of flatk as a whole. This is a
reasonable tradeoff for safety.